requirements.txt
anyio==4.7.0
boto3==1.35.76
botocore==1.35.76
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
databricks-cli==0.18.0
exceptiongroup==1.2.2
Faker==33.1.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jmespath==1.0.1
numpy==2.0.2
oauthlib==3.2.2
pandas==2.2.3
py4j==0.10.9.7
pyactiveresource==2.2.2
PyJWT==2.10.1
pyspark==3.5.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
s3transfer==0.10.4
schedule==1.2.2
ShopifyAPI==12.7.0
six==1.17.0
sniffio==1.3.1
stripe==11.3.0
tabulate==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==1.26.20
